+++ doc/Blamantic.Shared/Pages/Data/PaginationPage.razor
@@ -58,14 +58,14 @@
 ")
     </CodeTemplate>
 </PresentationPart>
-        <PresentationPart Title="Hide Total Count">
+        <PresentationPart Title="Hide Statistic">
     <RunTemplate>
-        <Pagination CurrentPage="1" TotalCount="100"  ShowTotalCount="false"/>
+        <Pagination CurrentPage="1" TotalCount="100"  ShowStatistic="false"/>
     </RunTemplate>
     <CodeTemplate>
         @Code.GetCode(@"
 ```html
-<Pagination CurrentPage=""1"" TotalCount=""100""  ShowTotalCount=""false""/>
+<Pagination CurrentPage=""1"" TotalCount=""100""  ShowStatistic=""false""/>
 ```
 ")
     </CodeTemplate>
