+++ src/Blamantic/Components/GridView/GridViewPaginationEventArgs.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace BlamanticUI
+{
+    public class GridViewPaginationEventArgs:EventArgs
+    {
+        public GridViewPaginationEventArgs(int currentPage,int pageSize)
+        {
+            CurrentPage = currentPage;
+            PageSize = pageSize;
+        }
+
+        public int CurrentPage { get; }
+        public int PageSize { get; }
+    }
+}
