+++ src/Blamantic/Components/GridView/GridViewTemplateBase.cs
@@ -26,7 +26,7 @@ namespace BlamanticUI
         /// <summary>
         /// Gets or sets the cascading value of data grid.
         /// </summary>
-        [CascadingParameter] protected GridView CascadingDataGrid { get; set; }
+        [CascadingParameter] protected GridView CascadingGridView { get; set; }
 
         /// <summary>
         /// Method invoked when the component is ready to start, having received its
@@ -35,7 +35,7 @@ namespace BlamanticUI
         /// <exception cref="InvalidOperationException">The '{GetType().Name}' must render inside of '{typeof(DataGrid).Name}'</exception>
         protected override void OnInitialized()
         {
-            if (CascadingDataGrid == null)
+            if (CascadingGridView == null)
             {
                 throw new InvalidOperationException($"The '{GetType().Name}' must render inside of '{typeof(GridView).Name}'");
             }
